refactoring to clean up
