+++ Sources/Chart/Marks/AnyMark.swift
@@ -12,27 +12,17 @@ import Foundation
 public struct AnyMark: Mark {
     public var xPropertyScale: VisualPropertyScale
     public var yPropertyScale: VisualPropertyScale
+    public var _yAxis: Axis?
+    public var _xAxis: Axis?
 
     private let wrappedSymbolsForMark: (_: CGRect) -> [MarkSymbol]
     private let wrappedAxisForMark: (_: CGRect) -> [Axis]
-    private let wrappedXAxis: () -> Axis?
-    private let wrappedYAxis: () -> Axis?
 
     public init<T: Mark>(_ specificMark: T) {
         wrappedSymbolsForMark = specificMark.symbolsForMark(in:)
         xPropertyScale = specificMark.xPropertyScale
         yPropertyScale = specificMark.yPropertyScale
         wrappedAxisForMark = specificMark.axisForMark(in:)
-        wrappedXAxis = specificMark.getXAxis
-        wrappedYAxis = specificMark.getYAxis
-    }
-
-    public func getXAxis() -> Axis? {
-        wrappedXAxis()
-    }
-
-    public func getYAxis() -> Axis? {
-        wrappedYAxis()
     }
 
     public func axisForMark(in rect: CGRect) -> [Axis] {
